feat: GetQuadrinhos funcionando perfeitamente

+++ server/node/quadrinho.js
@@ -1,8 +1,10 @@
 const express = require('express');
 const mysql = require('mysql2');
 const bodyParser = require('body-parser');
+const sharp = require('sharp');
 const app = express();
 const fs = require('fs');
+const path = require('path');
 
 const open = mysql.createConnection({
     host: 'localhost',
@@ -14,40 +16,65 @@ const open = mysql.createConnection({
 app.use(bodyParser.urlencoded({extended: false}));
 app.use(bodyParser.json());
 
-function createQuadrinho(req, res){
-    const camposObrigatorios = ['nome', 'valor', 'quantDispo', 'descricao', 'tipo', 'quantPaginas'];
-    const camposFaltando = [];
-    camposObrigatorios.forEach(field => {
-        if (!req.body[field]) {
-            camposFaltando.push(field);
-        }
-    });
 
-    if(req.file == undefined){
-        camposFaltando.push('foto');
-    }
 
-    if (camposFaltando.length > 0) {
-        fs.unlinkSync(req.file.path);
-        return res.status(400).json({ ok: false, camposObrigatorios: `${camposFaltando.join(', ')}` });
-    }else{
-        const sql = `INSERT INTO quadrinho(nome, valor, quantDispo, descricao, tipo, quantPaginas, foto) VALUES(?, ?, ?, ?, ?, ?, ?);`
-        const values = [req.body.nome, req.body.valor, req.body.quantDispo, req.body.descricao, req.body.tipo, req.body.quantPaginas, req.file.path];
-         open.query(sql, values, (err, result) => {
-            console.log(result);
-            if(err){
-                console.log(err);
-                res.send({ok: false, message: 'Erro ao inserir dados'})
-            }else{
-                res.send({ok: true})
+
+
+    function createQuadrinho(req, res){
+        const camposObrigatorios = ['nome', 'valor', 'quantDispo', 'descricao', 'tipo', 'quantPaginas'];
+        const camposFaltando = [];
+        camposObrigatorios.forEach(field => {
+            if (!req.body[field]) {
+                camposFaltando.push(field);
             }
-         })
-    }
+        });
     
-}
+        if(req.file == undefined){
+            camposFaltando.push('foto');
+        }
+    
+        if (camposFaltando.length > 0) {
+            if (req.file) {
+                fs.unlinkSync(req.file.path);
+            }
+            return res.status(400).json({ ok: false, camposObrigatorios: `${camposFaltando.join(', ')}` });
+        }else{
+            // Criar um novo caminho de arquivo para a imagem processada
+            const newFilePath = path.join(path.dirname(req.file.path), 'processed-' + path.basename(req.file.path));
+    
+            sharp(req.file.path)
+                .resize(186, 294)
+                .toFile(newFilePath, (err) => {
+                    if (err) {
+                        return res.status(500).json({ ok: false, message: 'Erro ao processar a imagem' });
+                    }
+    
+                    // Substituir o arquivo original pela imagem processada
+                    fs.rename(newFilePath, req.file.path, (err) => {
+                        if (err) {
+                            return res.status(500).json({ ok: false, message: 'Erro ao salvar a imagem' });
+                        }
+    
+                        const sql = `INSERT INTO quadrinho(nome, valor, quantDispo, descricao, tipo, quantPaginas, foto) VALUES(?, ?, ?, ?, ?, ?, ?);`
+                        const values = [req.body.nome, req.body.valor, req.body.quantDispo, req.body.descricao, req.body.tipo, req.body.quantPaginas, req.file.path];
+                        open.query(sql, values, (err, result) => {
+                            console.log(result);
+                            if(err){
+                                console.log(err);
+                                res.send({ok: false, message: 'Erro ao inserir dados'})
+                            }else{
+                                res.send({ok: true})
+                            }
+                        })
+                    });
+                });
+        }
+    
+    }
+
 
 function getQuadrinhos(req, res){
-    const sql = 'SELECT * FROM quadrinho;'
+    const sql = 'SELECT nome, foto, tipo, valor FROM quadrinho;'
     open.query(sql, (err, result) => {
         if (err){
             res.send({ok: false, message: 'Erro ao retornar quadrinhos'})
